# swagger.json
{
...
  "info": {
    "description": "Comprehensive API for bus payment system with offers, rewards, bookings, and refunds",
    "title": "Bus Payments & Rewards API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/customers": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "per_page",
            "schema": {
              "default": 10,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "email",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "current_page": {
                      "type": "integer"
                    },
                    "customers": {
                      "items": {
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "pages": {
                      "type": "integer"
                    },
                    "per_page": {
                      "type": "integer"
                    },
                    "total": {
                      "type": "integer"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "List of customers"
          }
        },
        "summary": "List all customers",
        "tags": [
          "Customers"
        ]
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "address": {
                    "type": "string"
                  },
                  "date_of_birth": {
                    "format": "date",
                    "type": "string"
                  },
                  "email": {
                    "format": "email",
                    "type": "string"
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  }
                },
                "required": [
                  "first_name",
                  "last_name",
                  "email"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Customer created successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "409": {
            "description": "Email already exists"
          }
        },
        "summary": "Create new customer",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/customers/{customer_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer deleted successfully"
          },
          "404": {
            "description": "Customer not found"
          }
        },
        "summary": "Delete customer",
        "tags": [
          "Customers"
        ]
      },
      "get": {
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer details"
          },
          "404": {
            "description": "Customer not found"
          }
        },
        "summary": "Get customer details",
        "tags": [
          "Customers"
        ]
      },
      "put": {
        "parameters": [
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "address": {
                    "type": "string"
                  },
                  "date_of_birth": {
                    "format": "date",
                    "type": "string"
                  },
                  "email": {
                    "format": "email",
                    "type": "string"
                  },
                  "first_name": {
                    "type": "string"
                  },
                  "last_name": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Customer updated successfully"
          },
          "400": {
            "description": "Invalid data"
          },
          "404": {
            "description": "Customer not found"
          }
        },
        "summary": "Update customer",
        "tags": [
          "Customers"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local development server",
      "url": "http://192.168.86.189:5001"
    }
  ]
}